refactor(page): extract SceneOverlays organism (phase 2)
Move the renderSceneOverlays render-prop body (hole-punch markers/gizmo/
preview cylinder, hollow voxel edit overlay + cavity/infill previews, the
LysGhost overlay) and its module-level helpers (transformVoxelCentersToWorld,
WorldSpaceVoxelPreview, WorldSpaceVoxelEditOverlay) into
components/organisms/scene/SceneOverlays.tsx. The SceneCanvas render prop
now delegates to <SceneOverlays ...sameNamedProps />; 5 now-orphaned
imports were pruned from page.tsx. Worktree subagent, integrated + verified.

page.tsx: 12,702 -> 12,431 lines (-271). Verified via `tsc --noEmit`.
Smoke-test hole-punch/hollow overlays in the 3D scene before merge.
